Updated all files for increased clarity and readability

+++ string/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: jde-baai <jde-baai@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/06 13:09:48 by jde-baai      #+#    #+#                 */
-/*   Updated: 2023/09/25 20:16:27 by juliusdebaa   ########   odam.nl         */
+/*   Updated: 2023/07/28 20:14:31 by juliusdebaa   ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,12 +17,11 @@
  * @param	s	String to calculate the length of
  * @return	Length of the string
  */
-size_t	ft_strlen(const char *s)
-{
-	size_t	i;
+size_t	ft_strlen(const char *s) {
+  unsigned int i;
 
-	i = 0;
-	while (s[i] != '\0')
-		i++;
-	return (i);
+  i = 0;
+  while (s[i] != '\0')
+    i++;
+  return (i);
 }
